+++ client/src/wallet/passkey/createCredential.ts
@@ -44,7 +44,10 @@ export const createPasskeyCredential = async ({
         name: userName,
         displayName: userName,
       },
-      pubKeyCredParams: [{ alg: -7, type: 'public-key' }],
+      pubKeyCredParams: [
+        { alg: -7, type: 'public-key' },
+        { alg: -257, type: 'public-key' },
+      ],
       authenticatorSelection: {
         authenticatorAttachment: 'platform',
         residentKey: 'required',